moved the box component into ui
