--- src/components/RepoScreen/index.tsx
@@ -1,5 +1,0 @@
-import { Text } from '../ui'
-
-export const RepoScreen = () => {
-  return <Text>Repo screen</Text>
-}
