+++ src/components/modals/layout.tsx
@@ -1,8 +1,7 @@
 import { ArrowLeft } from '@phosphor-icons/react'
 import { ReactNode } from 'react'
 
-import { Box } from '../BackupScreen/BackupDetail'
-import { ModalProps, Stack, Text } from '../ui'
+import { Box, ModalProps, Stack, Text } from '../ui'
 
 interface SharedModalLayoutProps extends Pick<ModalProps, 'onClose'> {
   title: string
