+++ src/components/ui/boxes.tsx
@@ -1,5 +1,7 @@
 import { css, styled } from 'next-yak'
 
+import { StyleProps } from './style'
+
 export const roundRectStyle = css`
   border-radius: 0.75rem;
   padding: calc(1rem - 1px);
@@ -10,4 +12,28 @@ export const roundRectStyle = css`
 
 export const Container = styled.div`
   padding: 3.4rem 3.2em;
+`
+
+export const Box = styled.div<Partial<StyleProps & { $isFocused?: boolean }>>`
+  border: ${({ $borderWidth = '1px', $isFocused }) =>
+      $isFocused ? '2px' : $borderWidth}
+    ${({ $borderStyle = 'dashed', $isFocused }) =>
+      $isFocused ? 'solid' : $borderStyle}
+    ${({ $borderColor = 'var(--color-gray-light)', $isFocused }) =>
+      $isFocused ? 'var(--color-dark-blue)' : $borderColor};
+  border-radius: 12px;
+  height: ${({ $height = '66px' }) => $height};
+  width: ${({ $width = '100%' }) => $width};
+  display: ${({ $display = '' }) => $display};
+  justify-content: ${({ $justifyContent = '' }) => $justifyContent};
+  align-items: center;
+  padding: ${({ $padding = '0 0.6rem' }) => $padding};
+  gap: ${({ $gap = 0 }) => $gap};
+  cursor: pointer;
+  background: ${({ $background = '' }) => $background};
+  position: ${({ $position = '' }) => $position};
+  top: ${({ $top = '' }) => $top};
+  right: ${({ $right = '' }) => $right};
+  left: ${({ $left = '' }) => $left};
+  bottom: ${({ $bottom = '' }) => $bottom};
 `
